apps/services/rpt-verify/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
PyNaCl==1.5.0
psycopg[binary]==3.2.1
